Moved ClientIteratorEnvironment out of API package, implement suggestions
